backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "FastAPI backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "sqlmodel>=0.0.22",
    "alembic>=1.13.0",
    "psycopg[binary]>=3.2.0",
    "pyjwt>=2.9.0",
    "pwdlib[argon2,bcrypt]>=0.2.0",
    "pydantic-settings>=2.6.0",
    "sentry-sdk[fastapi]>=2.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "coverage>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["app/tests"]
